Use floating point instead of integer division to calculate touch input display scale

+++ src/Xna.Framework.Input/Input/Touch/TouchPanelStrategy.Legacy.cs
@@ -175,8 +175,8 @@ namespace Microsoft.Xna.Platform.Input.Touch
             }
 
             // scale position
-            position.X = position.X * (DisplayWidth / winSize.X);
-            position.Y = position.Y * (DisplayHeight / winSize.Y);
+            position.X = position.X * ((float)DisplayWidth / winSize.X);
+            position.Y = position.Y * ((float)DisplayHeight / winSize.Y);
 
             TimeSpan currentTimestamp = this.CurrentTimestamp;
             int currentFramestamp = this._currentFramestamp;
